requirements.txt
Django==4.0.6
psycopg2>=2.8
django-mdeditor==0.1.20
django-preferences==1.0.0
django-imagekit==4.1.0
pillow==9.2.0
markdown==3.4.1
Pygments==2.13.0
django-extensions==3.2.0
requests==2.28.1